chatbot/requirements.txt
python-telegram-bot
python-dotenv
supabase
groq